fix mismatched label assignment between `real_loop` and `pattern_type` in the geographical pattern causality sources

+++ src/PatternCausality.cpp
@@ -328,7 +328,7 @@ PatternCausalityRes GenPatternCausality(
     //     continue;
     // }
 
-    res.RealLoop.push_back(t);
+    res.RealLoop.push_back(static_cast<int>(t));
     size_t i = pattern_indices[pat_x];
     size_t j = pattern_indices[pat_y_pred];
